chore: Remove Agent builder and ToolCallingAgent implementations
