+++ plugins/enthusiast-common/enthusiast_common/agents/__init__.py
@@ -1,2 +1,1 @@
 from .base import BaseAgent  # noqa
-from .base import ToolCallingAgent  # noqa
