+++ plugins/enthusiast-common/enthusiast_common/agents/base.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import Self
 
-from langchain.agents import AgentExecutor, create_tool_calling_agent
+from langchain.agents import AgentExecutor
 from langchain.memory import ConversationSummaryBufferMemory
 from langchain_core.language_models import BaseLanguageModel
 from langchain_core.prompts import ChatPromptTemplate
@@ -20,13 +20,20 @@ class BaseAgent(ABC):
         prompt: ChatPromptTemplate,
         conversation_service: BaseConversationService,
         conversation_id: int,
+        agent_executor: AgentExecutor,
+        memory: ConversationSummaryBufferMemory,
     ):
         self._tools = tools
         self._llm = llm
         self._prompt = prompt
         self._conversation_service = conversation_service
         self._conversation_id = conversation_id
-        self._memory = self._create_agent_memory(conversation_service.get_messages(self._conversation_id))
+        self._memory = memory
+        self._agent_executor = agent_executor
+
+    @abstractmethod
+    def get_answer(self, input_text: str) -> str:
+        pass
 
     @abstractmethod
     def _create_agent_executor(self, **kwargs) -> AgentExecutor:
@@ -43,20 +50,3 @@ class BaseAgent(ABC):
         agent_executor = self._create_agent_executor(**kwargs)
         self._set_agent_executor(agent_executor)
         return self
-
-
-class ToolCallingAgent(BaseAgent):
-    def _create_agent_executor(self, **kwargs):
-        agent = create_tool_calling_agent(self._llm, self._tools, self._prompt)
-        return AgentExecutor(agent=agent, tools=self._tools, verbose=True, **kwargs)
-
-    def get_answer(self, input_text: str) -> str:
-        agent_output = self._agent_executor.invoke({"input": input_text, "chat_history": self._memory.buffer})
-        return agent_output["output"]
-
-    def _create_agent_memory(self, messages) -> ConversationSummaryBufferMemory:
-        memory = ConversationSummaryBufferMemory(
-            llm=self._llm, memory_key="chat_history", return_messages=True, max_token_limit=3000, output_key="output"
-        )
-        memory.chat_memory.add_messages(messages)
-        return memory
