+++ plugins/enthusiast-common/enthusiast_common/builder/__init__.py
@@ -1,0 +1,2 @@
+from .base import BaseAgentBuilder  # noqa
+from .base import RepositoriesInstances  # noqa
